requirements.txt
discord.py
nest-asyncio
python-dotenv
openai